[step3] refactor: 테스트케이스 수정
