+++ src/test/java/learn/StubSystemInputLearnTest.java
@@ -1,8 +1,8 @@
-package util;
+package learn;
 
-import java.io.*;
+import java.io.ByteArrayInputStream;
 
-public class SystemInputStubUtil {
+public class StubSystemInputLearnTest {
 
     public static void stubInputStream(String input) {
